+++ src/main/java/com/fjgonmir/lil/learningspring/LearningSpringApplication.java
@@ -1,13 +1,7 @@
 package com.fjgonmir.lil.learningspring;
 
-import com.fjgonmir.lil.learningspring.entity.Room;
-import com.fjgonmir.lil.learningspring.repository.RoomRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
 
 @SpringBootApplication
 public class LearningSpringApplication {
