+++ src/main/java/com/fjgonmir/lil/learningspring/controller/web/RoomReservationController.java
@@ -1,7 +1,6 @@
 package com.fjgonmir.lil.learningspring.controller.web;
 
-import com.fjgonmir.lil.learningspring.Business.domain.RoomReservation;
-import com.fjgonmir.lil.learningspring.Business.services.ReservationService;
+import com.fjgonmir.lil.learningspring.business.services.ReservationService;
 import com.fjgonmir.lil.learningspring.utils.DateUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -11,7 +10,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import java.util.Date;
-import java.util.List;
 
 @Controller
 @RequestMapping("/room-reservations")
