+++ src/main/resources/static/js/site.js
@@ -7,11 +7,11 @@ function setPicker(){
                 }
             }
         });
-};
+}
 
 function getRequestParam(p){
-    return (window.location.search.match(new RegExp('[?&]' + p + '=([^&]+)')) || [, null])[1];
-};
+    return (window.location.search.match(new RegExp('[?&]' + p + '=([^&]+)')) || [null])[1];
+}
 
 function setInitialDate(){
     var requestDate = getRequestParam('date');
@@ -22,14 +22,14 @@ function setInitialDate(){
     }
     $('#datepicker').datepicker('setDate', requestDate);
 
-};
+}
 
 function reloadPageForDateSelection(){
     var selectedDate = document.getElementById('datepicker').value;
     var redirectLink = window.location.protocol + "//" + window.location.host + window.location.pathname + '?date=' + selectedDate;
     console.log('Redirecting to: ' + redirectLink);
     window.location.href = redirectLink;
-};
+}
 
 function formatDate(input) {
     var dateFormat = 'yyyy-mm-dd';
@@ -38,7 +38,7 @@ function formatDate(input) {
     dateFormat.replace(/(yyyy|dd|mm)/g, function(part) { fmt[part] = i++; });
 
     return new Date(parts[fmt['yyyy']], parts[fmt['mm']]-1, parts[fmt['dd']]);
-};
+}
 
 $(document).ready(function(){
     setPicker();
