+++ src/main/resources/templates/guests.html
@@ -34,13 +34,13 @@
         <input type="text" class="form-control" id="datepicker" aria-describedby="basic-addon1"/>
     </div>
 
-    <table class="table table-striped table-bordered">
-        <tr>
-            <td>first Name</td>
-            <td>Last Name</td>
-            <td>phone number</td>
-            <td>mail address</td>
-        </tr>
+    <table class="table table-striped table-bordered" aria-describedby="table for show guest info">
+        <thead>
+            <th scope="col">first Name</th>
+            <th scope="col">Last Name</th>
+            <th scope="col">phone number</th>
+            <th scope="col">mail address</th>
+        </thead>
         <tr th:each="guest: ${guestList}">
             <td th:text="${guest.firstName}!=null? ${guest.firstName} : 'VACANT'">first Name</td>
             <td th:text="${guest.lastName}!=null? ${guest.lastName} : 'VACANT'">last Name</td>
