+++ src/main/resources/templates/reservations.html
@@ -33,12 +33,12 @@
         <input type="text" class="form-control" id="datepicker" aria-describedby="basic-addon1"/>
     </div>
 
-    <table class="table table-striped table-bordered">
-        <tr>
-            <td>Room</td>
-            <td>Room Number</td>
-            <td>Guest</td>
-        </tr>
+    <table class="table table-striped table-bordered"  aria-describedby="table for show room reservations info">>
+        <thead>
+            <th scope="col">Room</th>
+            <th scope="col">Room Number</th>
+            <th scope="col">Guest</th>
+        </thead>
         <tr th:each="roomReservation: ${roomReservations}">
             <td th:text="${roomReservation.roomName}">Room Name</td>
             <td th:text="${roomReservation.roomNumber}">Room Number</td>
